add all state variables into constructor
It is valuable if all state variables, or properties of an object, are in the
constructor. This is is to facilitate easily understanding what the state of an
object can be, and producing an object with the desired state. It is also the
same convention used throughout Fullnode and Yours.

This class has an extremely complicated state with many more properties than
anything else in Fullnode or Yours, which is a sign that it should be broken up
into smaller conceptual pieces. Objects with complicated states are harder to
understand and harder to debug.