fix random numbers
Math.random() is not cryptographically secure and should never be used in a
cryptographic context. Random.getRandomBuffer, from Fullnode, uses the
browserify libraries that are based on the operating system's entropy pool in
both node.js and browsers. 32 bytes (256 bits) is used, because that is the
standard for unbreakable by anyone for the foreseeable future. Even 128 bits is
considered secure; 256 bits goes further.