spend commitment tx and change scripts
There was a minor problem with the scripts as laid out in yours-lightning.md.
Because all CHECKSIGVERIFY and EQUALVERIFY were used, these actually drop the
top item on the stack, leaving the stack empty at the end of execution, making
the script invalid. The most efficient solution is simply to replace the last
"XXXVERIFY" in each execution path with "XXX" (no "VERIFY").

I have verified only one spend - the spending of the normal HTLC output by Bob
who knows the HTLC secret, because Alice has chosen to settle the transaction.
There are other execution paths I have not yet tested - there may be more bugs
in the script I have not uncovered.