Msg and MsgOpenChannel
We will need the ability to build and send messages from channel agent to
channel agent. These messages will specify what actions to perform for the
other agent. The first most obvious message to implement is "Open Channel".
i.e., imagine Bob wants to open a channel with Carol. The first message he
sends to her is the Open Channel message. This message will need to contain the
public key for their shared multisig address. It also contains the quantity of
bitcoins (denominated in satoshis, in the form of a big number) that he intends
to fund the channel with. Carol must agree that funding the channel with that
quantity is acceptable.