chanId
It has occurred to me that it will be necessary for channels to have an ID,
because it could be that two agents may wish to have multiple channels open
with each other. At first I thought the ID could simply be the funding multisig
address which uniquely identifies the channel, but unfortunately the multisig
address is not generated until after the channel already exists. So there needs
to be some kind of ID generated by the channel initiator. For now it just
defaults to a random string.