use conditional shorthand
Using the ? : syntax for conditionals reduces the code size to approximately
1/3. Technically, there is a subtle difference - rather than not set the key,
we are setting the key to undefined, which means Object.keys() returns
something different. But this will not be noticeable most of the time, and has
less code and is easier to understand, and is consistent with the rest of our
codebase.