MsgUpdateRes
An update message should be followed by an update-res message, not just another
update message. This reason for this is because both agents may send each other
an update message at the same time, causing a conflict. The solution is to
require a different message type as a response, so you know whether they are
trying to send you a response or whether it is a new update message.