initial work on "channel"
agent.js requires memory access to the other agent, which won't work in a
real-world environment. We need a new class based on sending and receiving
messages that doesn't assume memory access to the other agent. I have begun
such a thing in channel.js. Instead of creating an "agent", you will create a
"channel". It is not done - I'm not sure I have the protocol or required
messages right yet. I have written up an explanation about how to use this new
class (assuming it were finished) in docs/yours-lightning.md.