notion of "conversation id" for messages
Often, the Yours Lightning Protocol requires messages to be sent in a
particular sequence, each one expecting a particular response from the other
agent. Together, these messages constitute a "conversation". In order to
validate that you are seeing messages for the right conversation, and not
another conversation initiated by the other agent, each message in the same
conversation needs to have the same convId, or "conversation id".