simplify low-level protocol
We've realized that because Bob and Carol already know each other's public
keys, this can reduce the number of messages that need to be sent. Opening a
channel now merely requires sending the first update message for that channel.
We have included a graphic of a state machine for how the message protocol
works. Our protocol documentation does not currently have information on how to
proceed in the case of errors.