add state transitions
We're using a finite automata model for the states of channel which are
currently sketched out in state-machine.js. I have updated channel.js to
reflect these state transitions. This will help with understanding the code and
with debugging. The actual logic of what to do with messages or how to build
them is not yet in channel.js.