merge old protocol spec into new protocol spec
We actually had three protocol specs in the document. One was old, one was new,
and one was a very detailed one. I've merged the old one into the new one and
left the detailed one alone for now. I've then also copied the old one into
agent.js since it is premature to remove it - perhaps it contains details that
I have accidentally left unmerged.