remove msgUpdateRes
In the latest version of our state machine for the message handler,
msgUpdateRes (or 'update-res', the response to an 'update' message), does not
exist, because you can simply send a msgUpdate in response to a msgUpdate. So I
have removed the obsolete msgUpdateRes code and references to it in the
documentation.