one more spend
There was a bug in spending.asyncBuild where it was adding a number of
inputs that may be less than the length of the output list, but where
when it went to sign the inputs, it assumed the length was the same. It
is necessary to keep track of the number of inputs separately, since you
may not be able to spend every output in the output list. I keep track
of this with the standard nIn variable which is what is used in Bitcoin.